Move last functions to shared.functions.
