+++ components/server/src/external/database/measurements.py
@@ -8,7 +8,7 @@ from pymongo.database import Database
 
 from shared.model.measurement import Measurement
 from shared.model.metric import Metric
-from server_utilities.functions import iso_timestamp
+from shared.utils.functions import iso_timestamp
 from server_utilities.type import MetricId
