+++ components/server/src/external/database/reports.py
@@ -7,8 +7,7 @@ from pymongo.database import Database
 
 from shared.database.filters import DOES_EXIST, DOES_NOT_EXIST
 from shared.model.report import Report
-
-from server_utilities.functions import iso_timestamp
+from shared.utils.functions import iso_timestamp
 from server_utilities.type import MetricId, ReportId, SubjectId
 
 from ..database import sessions
