+++ components/server/src/shared/model/source.py
@@ -6,7 +6,7 @@ from collections.abc import Sequence
 from datetime import datetime
 from typing import cast, Optional, TYPE_CHECKING
 
-from server_utilities.functions import days_ago, iso_timestamp
+from shared.utils.functions import days_ago, iso_timestamp
 
 
 if TYPE_CHECKING:
