+++ components/server/src/shared/utils/functions.py
